requirements.txt
azure-openai
streamlit
PyPDF2
python-dotenv
